fixed range to type float

+++ src/modules/localization/include/segmenter/segment_processor.h
@@ -10,7 +10,7 @@ class SegmentProcessor
 
         void CalculateBoundingBox(); //TODO
 
-        int PointsInRange(size_t cluster_index, int x, int y);
+        int PointsInRange(size_t cluster_index, float x, float y);
 
         pcl::PointXYZI CalculateCentroid(int cluster_index);
